Trilinear form and example updates.
Standardized the signs for trilinear form implementation in the direct solver so
it is 100% clear which form is being used.

Added some new examples and changed the existing so that they are sufficiently
distinguished from one another (i.e. boundary conditions, load, etc.).

Added in function printMuComparison that tries to generate output graphs in a
form suitable for printing (hi-res TIFF).